feat(admin-2-input): add basic component
