--- app/src/views/SimplifiedEapForm/EarlyAction/DistrictMap/DistrictMapModal/DistrictItem/i18n.json
@@ -1,7 +1,0 @@
-{
-    "namespace": "simplifiedEapForm",
-    "strings": {
-        "removeDistrict": "Remove Districts",
-        "removeAdmin2":"Remove Admin2"
-    }
-}
