--- app/src/views/SimplifiedEapForm/EarlyAction/DistrictMap/DistrictMapModal/DistrictItem/styles.module.css
@@ -1,56 +1,0 @@
-.district-item {
-    display: flex;
-    flex-direction: column;
-    background-color: var(--go-ui-color-background);
-    padding: var(--go-ui-spacing-xs) 0;
-
-    .district {
-        display: flex;
-        padding: var(--go-ui-spacing-xs) var(--go-ui-spacing-md);
-        font-weight: var(--go-ui-font-weight-medium);
-
-        .name {
-            flex-grow: 1;
-        }
-
-        .remove-button {
-            flex-shrink: 0;
-            opacity: 0;
-        }
-
-        &:hover {
-            background-color: var(--go-ui-color-background-hover);
-            .remove-button {
-                opacity: 1;
-            }
-        }
-    }
-
-    .admin2-items {
-        display: flex;
-        flex-direction: column;
-
-        .admin-2-item {
-            display: flex;
-            padding: var(--go-ui-spacing-3xs) var(--go-ui-spacing-md);
-            font-size: var(--go-ui-font-size-sm);
-
-            .name {
-                flex-grow: 1;
-            }
-
-            .remove-button {
-                flex-shrink: 0;
-                opacity: 0;
-            }
-
-            &:hover {
-                background-color: var(--go-ui-color-background-hover);
-
-                .remove-button {
-                    opacity: 1;
-                }
-            }
-        }
-    }
-}
