--- app/src/views/SimplifiedEapForm/EarlyAction/DistrictMap/DistrictMapModal/i18n.json
@@ -1,11 +1,0 @@
-{
-    "namespace": "simplifiedEapForm",
-    "strings": {
-        "districtButtonSelectProvince": "Select Province / Region",
-        "save": "Save",
-        "backToCountry": "Back to Country",
-        "labelDistricts": "Districts",
-        "listEmptyMessage":"Select districts by click on map or from the select input above",
-        "mapTitle": "District / Admin2 selection map"
-    }
-}
