--- app/src/views/SimplifiedEapForm/EarlyAction/DistrictMap/DistrictMapModal/styles.module.css
@@ -1,36 +1,0 @@
-.district-map-modal {
-    .content {
-        display: flex;
-        gap: var(--go-ui-spacing-sm);
-
-        .map-section {
-            display: flex;
-            flex-direction: column;
-            gap: var(--go-ui-spacing-sm);
-
-            .map-container {
-                width: 40rem;
-                height: 60vh;
-            }
-
-            .back-button {
-                font-size: var(--go-ui-height-icon-multiplier);
-            }
-        }
-
-        .side-panel {
-            display: flex;
-            flex-direction: column;
-            gap: var(--go-ui-spacing-sm);
-            height: 60vh;
-
-            .district-list {
-                display: flex;
-                flex-direction: column;
-                overflow-y: auto;
-                gap: var(--go-ui-spacing-xs);
-            }
-        }
-
-    }
-}
