--- app/src/views/SimplifiedEapForm/EarlyAction/DistrictMap/i18n.json
@@ -1,6 +1,0 @@
-{
-    "namespace": "simplifiedEapForm",
-    "strings": {
-        "buttonSelectProvince": "Select Province / Region"
-    }
-}
