--- app/src/views/SimplifiedEapForm/EarlyAction/DistrictMap/styles.module.css
@@ -1,14 +1,0 @@
-.input {
-    display: flex;
-    flex-direction: column;
-    gap: var(--go-ui-spacing-sm);
-
-    .district-list {
-        display: flex;
-        flex-direction: column;
-
-        .local-units {
-            font-size: var(--go-ui-font-size-sm);
-        }
-    }
-}
